# Conflicts:
#	app/build.gradle
#	app/src/main/java/com/example/madbudget/RecipeActivity.kt
#	app/src/main/java/com/example/madbudget/Recipes.kt
#	app/src/main/java/com/example/madbudget/models/Ingredient.kt
#	app/src/main/java/com/example/madbudget/models/Recipe.kt
#	app/src/main/java/com/example/madbudget/salling/ActivitySallingTest.kt
#	app/src/main/res/layout/activity_recipe.xml
#	app/src/main/res/layout/activity_recipes.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/example/madbudget/salling/SallingCommunicator.kt
@@ -19,7 +19,7 @@ class SallingCommunicator {
         //filter: zip, city, country ("dk", "se", "de", "pl"), street, brand ("netto", "bilka", "foetex", "salling", "carlsjr", "br")
         //fields: address, brand, coordinates, created, distance_km, hours, modified, name, phoneNumber, sapSiteId, type, vikingStoreId, attributes, id
 
-        fun getNearbyStores(context: Context, radius: Int, callback: Response.Listener<String>) {
+        fun getNearbyStores(context: Context, radius: Double, callback: Response.Listener<String>) {
 
             var fusedLocationClient = startLocationUpdates(context)
